refactor(auth): extract auth-store persistence shard
